Moved RobotPanel/RobotPanelHandler to library
